# openapi.yaml
openapi: 3.0.3
...
info:
  title: RTK Account Manager API
  version: 0.1.0
  description: |
    RTK Account Manager is a backend service API and authoritative account-side
    control plane for identity, tenant context, authorization, device registry,
    and provisioning intent.

    This service does not provide a WebUI, dashboard, frontend console, BFF
    session layer, UI preferences, or display-cache ownership. WebUI and
    dashboard experiences belong to client or console repositories such as
    `rtk_cloud_admin`; those clients consume this OpenAPI contract instead of
    being implemented in this server.
...
servers:
  - url: http://localhost:8080
paths:
...
  /v1/developer/brand-clouds/{brandCloudId}/members/{userId}/disable:
    patch:
      tags: [Developer Brand Clouds]
      operationId: disableDeveloperBrandCloudMember
      x-rtk-feature-id: FEAT-CA-BRAND-001
      x-rtk-requirement-ids:
        - REQ-UI-CA-CLOUD-005
      summary: Disable a member in this Brand Cloud without disabling the global account.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/BrandCloudId"
        - $ref: "#/components/parameters/UserId"
      responses:
        "200": { description: Member disabled. }
        "409": { $ref: "#/components/responses/Error" }
  /v1/developer/brand-clouds/{brandCloudId}/members/{userId}/enable:
    patch:
      tags: [Developer Brand Clouds]
      operationId: enableDeveloperBrandCloudMember
      x-rtk-feature-id: FEAT-CA-BRAND-001
      x-rtk-requirement-ids:
        - REQ-UI-CA-CLOUD-005
      summary: Re-enable a member in this Brand Cloud.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/BrandCloudId"
        - $ref: "#/components/parameters/UserId"
      responses:
        "200": { description: Member enabled. }
    delete:
      tags: [Developer Brand Clouds]
      operationId: removeDeveloperBrandCloudMember
      x-rtk-feature-id: FEAT-CA-BRAND-001
      x-rtk-requirement-ids:
        - REQ-UI-CA-CLOUD-005
      summary: Remove a Developer Brand Cloud member.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/BrandCloudId"
        - $ref: "#/components/parameters/UserId"
      responses:
        "204": { description: Member removed. }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
...
  /v1/admin/brand-clouds:
    post:
      tags:
        - Brand Clouds
      operationId: createBrandCloud
      x-rtk-feature-id: FEAT-CA-BRAND-001
      x-rtk-requirement-ids:
        - REQ-UI-CA-CLOUD-004
        - REQ-UI-CA-AUDIT-001
      security:
        - bearerAuth: []
      summary: Create a platform-managed brand cloud.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrandCloudRequest"
      responses:
        "201":
          description: Brand cloud created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandCloudResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
    get:
      tags:
        - Brand Clouds
      operationId: listBrandClouds
      x-rtk-feature-id: FEAT-CA-BRAND-001
      x-rtk-requirement-ids:
        - REQ-UI-CA-CLOUD-004
      security:
        - bearerAuth: []
      summary: List platform-managed brand clouds.
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Brand cloud list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandCloudsResponse"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /v1/admin/brand-clouds/{brandCloudId}:
    get:
      tags:
        - Brand Clouds
      operationId: getBrandCloud
      x-rtk-feature-id: FEAT-CA-BRAND-001
      x-rtk-requirement-ids:
        - REQ-UI-CA-CLOUD-004
      security:
        - bearerAuth: []
      summary: Get a platform-managed brand cloud.
      parameters:
        - $ref: "#/components/parameters/BrandCloudId"
      responses:
        "200":
          description: Brand cloud.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandCloudResponse"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      tags:
        - Brand Clouds
      operationId: updateBrandCloud
      x-rtk-feature-id: FEAT-CA-BRAND-001
      x-rtk-requirement-ids:
        - REQ-UI-CA-CLOUD-004
      security:
        - bearerAuth: []
      summary: Update a platform-managed brand cloud.
      parameters:
        - $ref: "#/components/parameters/BrandCloudId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrandCloudRequest"
      responses:
        "200":
          description: Brand cloud updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandCloudResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/admin/brand-clouds/{brandCloudId}/device-item-profiles:
    post:
      tags:
        - Brand Clouds
      operationId: createDeviceItemProfile
      x-rtk-feature-id: FEAT-CA-BRAND-001
      x-rtk-requirement-ids:
        - REQ-UI-CA-CLOUD-004
      security:
        - bearerAuth: []
      summary: Create a brand-cloud device item profile.
      parameters:
        - $ref: "#/components/parameters/BrandCloudId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceItemProfileRequest"
      responses:
        "201":
          description: Device item profile created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceItemProfileResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    get:
      tags:
        - Brand Clouds
      operationId: listDeviceItemProfiles
      x-rtk-feature-id: FEAT-CA-BRAND-001
      x-rtk-requirement-ids:
        - REQ-UI-CA-CLOUD-004
      security:
        - bearerAuth: []
      summary: List brand-cloud device item profiles.
      parameters:
        - $ref: "#/components/parameters/BrandCloudId"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/DeviceItemProfileStatus"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Device item profile list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceItemProfilesResponse"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/admin/brand-clouds/{brandCloudId}/device-item-profiles/{profileId}:
    get:
      tags:
        - Brand Clouds
      operationId: getDeviceItemProfile
      x-rtk-feature-id: FEAT-CA-BRAND-001
      x-rtk-requirement-ids:
        - REQ-UI-CA-CLOUD-004
      security:
        - bearerAuth: []
      summary: Get a brand-cloud device item profile.
      parameters:
        - $ref: "#/components/parameters/BrandCloudId"
        - $ref: "#/components/parameters/ProfileId"
      responses:
        "200":
          description: Device item profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceItemProfileResponse"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      tags:
        - Brand Clouds
      operationId: updateDeviceItemProfile
      x-rtk-feature-id: FEAT-CA-BRAND-001
      x-rtk-requirement-ids:
        - REQ-UI-CA-CLOUD-004
      security:
        - bearerAuth: []
      summary: Update a brand-cloud device item profile.
      parameters:
        - $ref: "#/components/parameters/BrandCloudId"
        - $ref: "#/components/parameters/ProfileId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceItemProfileRequest"
      responses:
        "200":
          description: Device item profile updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceItemProfileResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/admin/brand-clouds/{brandCloudId}/device-item-profiles/{profileId}/disable:
    post:
      tags:
        - Brand Clouds
      operationId: disableDeviceItemProfile
      x-rtk-feature-id: FEAT-CA-BRAND-001
      x-rtk-requirement-ids:
        - REQ-UI-CA-CLOUD-004
      security:
        - bearerAuth: []
      summary: Disable a brand-cloud device item profile.
      parameters:
        - $ref: "#/components/parameters/BrandCloudId"
        - $ref: "#/components/parameters/ProfileId"
      responses:
        "200":
          description: Device item profile disabled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceItemProfileResponse"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
  /v1/admin/brand-clouds/{brandCloudId}/users:
    get:
      tags:
        - Brand Clouds
      operationId: listBrandCloudAccounts
      x-rtk-feature-id: FEAT-CA-BRAND-001
      x-rtk-requirement-ids:
        - REQ-UI-CA-CLOUD-004
      security:
        - bearerAuth: []
      summary: List global users and memberships for one Brand Cloud.
      parameters:
        - $ref: "#/components/parameters/BrandCloudId"
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, pending_verification, disabled]
        - name: q
          in: query
          required: false
          schema:
            type: string
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Global user membership list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandCloudAccountsResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    post:
      tags:
        - Brand Clouds
      operationId: provisionBrandCloudAccountUser
      x-rtk-feature-id: FEAT-CA-BRAND-001
      x-rtk-requirement-ids:
        - REQ-UI-CA-CLOUD-004
      security:
        - bearerAuth: []
      summary: Find or create a global user and assign a Brand Cloud membership.
      description: "Owner provisioning always uses global email activation. `activation_mode=immediate` is restricted to audited staging creation of admin/member load-test accounts."
      parameters:
        - $ref: "#/components/parameters/BrandCloudId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrandCloudAccountProvisionRequest"
      responses:
        "200":
          description: Existing global user assigned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandCloudAccountProvisionResponse"
        "201":
          description: Global user created and assigned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandCloudAccountProvisionResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
  /v1/orgs/{orgId}/members/{userId}:
    patch:
      tags:
        - Members
      operationId: updateMemberRole
      x-rtk-feature-id: FEAT-AM-IDENTITY-001
      x-rtk-requirement-ids:
        - REQ-AM-MEMBERSHIP-INVARIANT-001
      security:
        - bearerAuth: []
      summary: Update member role.
      description: Refuses to downgrade the last active organization owner. Disabled owner memberships do not satisfy the active-owner invariant.
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMemberRequest"
      responses:
        "200":
          description: Member updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
    delete:
      tags:
        - Members
      operationId: removeMember
      x-rtk-feature-id: FEAT-AM-IDENTITY-001
      x-rtk-requirement-ids:
        - REQ-AM-MEMBERSHIP-INVARIANT-001
      security:
        - bearerAuth: []
      summary: Remove member from organization.
      description: Refuses to remove the last active organization owner. Disabled owner memberships do not satisfy the active-owner invariant.
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/UserId"
      responses:
        "204":
          description: Member removed.
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    BrandCloudId:
      name: brandCloudId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 200
    Offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            invalid_claim_token:
              summary: Non-retryable Claim Token error.
              value:
                error:
                  code: invalid_claim_token
                  message: Invalid claim token
                  retryable: false
                  resolution_action: scan_or_enter_a_valid_claim_token
            claim_service_unavailable:
              summary: Retryable Claim Token service error.
              value:
                error:
                  code: service_unavailable
                  message: Claim resolve is temporarily unavailable
                  retryable: true
                  resolution_action: retry_later
            oidc_user_not_provisioned:
              summary: Unknown SSO user.
              value:
                error:
                  code: user_not_provisioned
                  message: SSO user is not provisioned in account manager
            oidc_disabled:
              summary: OIDC login disabled.
              value:
                error:
                  code: oidc_disabled
                  message: OIDC authentication is disabled
            invalid_oidc_state:
              summary: Invalid or replayed OIDC state.
              value:
                error:
                  code: invalid_oidc_state
                  message: Invalid OIDC login state
            invalid_oidc_token:
              summary: Invalid OIDC token.
              value:
                error:
                  code: invalid_oidc_token
                  message: Invalid OIDC token
            unverified_oidc_email:
              summary: Unverified OIDC email.
              value:
                error:
                  code: unverified_oidc_email
                  message: OIDC email is not verified
            oidc_provider_misconfigured:
              summary: OIDC provider misconfigured.
              value:
                error:
                  code: oidc_provider_misconfigured
                  message: OIDC provider is misconfigured
  schemas:
...
    BrandCloudRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        tenant_slug:
          type: string
          minLength: 1
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
        status:
          $ref: "#/components/schemas/OrganizationStatus"
        metadata:
          type: object
          additionalProperties: true
...
    BrandCloudResponse:
      type: object
      required: [brand_cloud]
      properties:
        brand_cloud:
          $ref: "#/components/schemas/Organization"
...
    BrandCloudsResponse:
      type: object
      required: [brand_clouds, pagination]
      properties:
        brand_clouds:
          type: array
          items:
            $ref: "#/components/schemas/Organization"
        pagination:
          $ref: "#/components/schemas/Pagination"
        developer_cloud_limit:
          type: integer
          minimum: 1
...
    OrganizationStatus:
      type: string
      enum: [active, disabled]
...
    DeviceItemProfileStatus:
      type: string
      enum: [active, disabled]
...